[Refactor] 초기 데이터 설정 Custom hook으로 분리
